pyproject.toml
[tool.poetry]
name = "rag"
version = "0.1.0"
description = ""
authors = ["Michal K <michalkoniecznu0292@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
pydantic-ai = ">=0.1.11,<0.2.0"
crawl4ai = ">=0.6.2,<0.7.0"
psutil = ">=7.0.0,<8.0.0"
logfire = ">=3.15.0,<4.0.0"
google-genai = ">=1.14.0,<2.0.0"
supabase = ">=2.15.1,<3.0.0"
dotenv = ">=0.9.9,<0.10.0"
aiolimiter = ">=1.2.1,<2.0.0"
langchain-text-splitters = ">=0.3.8,<0.4.0"
pandas = ">=2.2.3,<3.0.0"
chromadb = "^1.0.9"
streamlit = "^1.45.1"
markdownify = "^1.1.0"
beautifulsoup4 = "^4.13.4"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[tool.logfire]
project_name = "rag-knowledge-base"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

